Expose base.py as a public api.
Extract methods that can be used to implement Websocket over HTTP/2

The implementation of HTTP/2 handshake is quite similar to HTTP/1 with a
few differences:
- headers are lower case in HTTP/2
- not exactly the same headers must be validated
- We don't handle Sec-WebSocket-Accept for HTTP/2
- the response is different

So we can implement the handshake in a base class that will delegate
some operations to the parent class.

We make _base.py public by renaming it to base.py
