+++ test/test_handshake.py
@@ -28,7 +28,7 @@
 # THEORY OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT
 # (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE
 # OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
-"""Tests for handshake._base module."""
+"""Tests for handshake.base module."""
 
 from __future__ import absolute_import
 import unittest
@@ -39,8 +39,8 @@ from mod_pywebsocket.common import ExtensionParameter
 from mod_pywebsocket.common import ExtensionParsingException
 from mod_pywebsocket.common import format_extensions
 from mod_pywebsocket.common import parse_extensions
-from mod_pywebsocket.handshake._base import HandshakeException
-from mod_pywebsocket.handshake._base import validate_subprotocol
+from mod_pywebsocket.handshake.base import HandshakeException
+from mod_pywebsocket.handshake.base import validate_subprotocol
 
 
 class ValidateSubprotocolTest(unittest.TestCase):
